+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.7...HEAD) - TBD
 
+### :wrench: Changed
+
+- **INTERNAL**: Use `str` in enums.
+
 ## [4.0.0-alpha.7](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.6...v4.0.0-alpha.7) - 2025-03-21
 
 ### :rocket: Added