+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Generate empty header values during the coverage phase.
 - Correctly handle second CTRL-C when waiting for worker threads.
 - Generate negative values for `minItems` & `maxItems` during the coverage phase.
+- Use `default` value as valid input during the coverage phase.
 
 ## [4.0.0-alpha.7](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.6...v4.0.0-alpha.7) - 2025-03-21