+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Generate negative values for `minItems` & `maxItems` during the coverage phase.
 - Use `default` value as valid input during the coverage phase.
 - Unknown `multipart/form-data` fields not added to the final test case payload.
+- Generate a non-empty string for negative testing of `type: string` enums during the coverage phase.
 
 ## [4.0.0-alpha.7](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.6...v4.0.0-alpha.7) - 2025-03-21